+++ packages/xcm-sdk/CHANGELOG.md
@@ -45,7 +45,8 @@
 
 ### Major Changes
 
-- Fee swaps & validations
+- Support xchain swaps
+- Support transfer validations
 
 ### Patch Changes